Enable EDAC debug facility on config-debug
It helps to have EDAC debug enabled, on -debug modules. the actual
debug only happens when a modprobe parameter is used, so the only
side effect of enabling the debug is to increase a little bit
the size of the edac Kernel modules, as new printk's will be
available there.

While here, bump version to 3.6.2-3.

Signed-off-by: Mauro Carvalho Chehab <mchehab@redhat.com>