Turn off W+X warnings (rhbz 1306885)
The W+X warning is useful for finding potentially insecure
mappings. Fixes for some of the warning have stalled leading
to lots of reports of warnings though. There isn't much use
for this in stable releases since any fix will be trickling
in through rawhide anyway. Turn it off for stable releases
and leave it on for rawhide.