Include version field to generated BLS configuration fragment
The BootLoaderSpec defines an optional version field that contains a human
readable version string for the menu item. This can be used by bootloaders
to set the boot menu item names instead of the optional title field.

For example the zipl bootloader used by s390x architecture doesn't support
names that contains spaces, so it uses the kernel version as the item name.

Signed-off-by: Javier Martinez Canillas <javierm@redhat.com>