Revert "Cleanup some binaries for kernel-devel (rhbz 1550152)"
This reverts commit 842f170f35de359b0c58707f8afca29d63ff631e.

It turns out that despite having all the source files available, the
kernel make system is not smart enough to automatically rebuild certain
things like modpost or recordmcount when building an external modules.
Really if we're going to have to copy everything under the sun so
'make prepare' succeeds it would be nice just to build things. But that
requires more discussion upstream so just revert this change for now.
Long live external modules.