Adjust grep pattern for finding Kconfig symbols
This script detects unknown Kconfig symbols by checking whether the
Kconfig macros used in the various *.config files have a corresponding
Kconfig symbol in the tree. Its core test is done with grep().

It turns out that match pattern used in grep() is too broad. For
instance, it doesn't report CONFIG_ACPI_PROCFS as unknown because there
is a Kconfig entry for ACPI_PROCFS_POWER in the tree. Make the pattern
match Kconfig symbols exactly to correct this.

Signed-off-by: Paul Bolle <pebolle@tiscali.nl>