Disable ISDN as upstream is moving to deprecate it
Some of the ISDN stack was deleted in the 5.3 window as per Paul's cleanup
patch. The vast majority of the remainder was moved to staging so let's
disable it in Fedora as it's unlikely to be used, and if se we can feed this
back to upstream.

For reference the main details from upstream are in these commits:
8a7e8ff8ce8a : ISDN removal/move merge overview
6d97985072dc : isdn: move capi drivers to staging
9c3c0c204814 : isdn: remove isdn4linux
85993b8c9786 : isdn: remove hisax driver