Drop cpumask auto select patch and set NR_CPUS appropriately
We've been carrying a patch to make CPUMASK_OFFSTACK selectable
without debugging for a long time now. The comment said this was
going to be replaced with something else but that never seemed
to happen. We're carrying it to have a higher number of CPUs but
at this point, adjusting NR_CPUS doesn't really get us that much
benfit. Drop the patch and just use 8192 or NR_CPUS on x86.

Signed-off-by: Laura Abbott <labbott@redhat.com>