Be smarter about getting rid of debuginfo
We've been doing a workaround of using a uuid for the build-id for
userspace binaries to let us get parallel debuginfo (the trick we
use for the kernel doesn't extend to userspace binaries). This
works but isn't ideal for reproducibility. Given we don't actually
care about this debuginfo, just remove it from the buildroot completely.