Resolve syntax warnings in SPEC file
The "regular" definition of the with_kabichk macro does not need to
be commented out while it is temporarily (re-)defined as 0 on the
next line. This avoids a "Macro expanded in comment" warning.

Remove comments on the same line as %endif directives. Instead, make
a couple of these conditionals easier to read (and consistent with
the surrounding ones) by deleting a few blank lines.