Disable CONFIG_DRM_DP_CEC temporarily
This setting introduces a dependency in drm_kms_helpers.ko on cec.ko but
cec.ko is in the media subsystem and is filtered into kernel-modules,
but drm_kms_helpers.ko is in kernel-core. There's not an easy way to
pick up cec.ko for kernel-core to satisfy the module dependency in the
current filtering scripts so I'm turning this off until the script can
be adjusted to make this easier.

When flipping this back on, CONFIG_CEC_CORE also needs to be turned on.